+++ src/main/scala/fastml4j/losses/Loss.scala
@@ -8,20 +8,20 @@ import org.nd4j.linalg.factory.Nd4j
 
 
 abstract class Loss{
-  protected var lambdaL2: Double = 0
-  protected var lambdaL1: Double = 0
+  protected var lambdaL2: Float = 0
+  protected var lambdaL1: Float = 0
 
-  def loss(weights: INDArray, dataSet: DataSet): Double
+  def loss(weights: INDArray, dataSet: DataSet): Float
   def gradient(weights: INDArray,dataSet: DataSet): INDArray
 
   // More about gradient checking:  http://cs231n.github.io/neural-networks-3/
-  def numericGradient(weights: INDArray, dataSet: DataSet, eps: Double = 1e-6): INDArray =
+  def numericGradient(weights: INDArray, dataSet: DataSet, eps: Float = 1e-6f): INDArray =
     (0 until weights.columns).map {
       i =>
         val oldWeights = weights.dup.put(0, i, weights.get(0,i) - eps)
         val newWeights = weights.dup.put(0, i, weights.get(0,i) + eps)
 
-        (loss(newWeights, dataSet) - loss(oldWeights, dataSet)) / 2.0 / eps }
+        (loss(newWeights, dataSet) - loss(oldWeights, dataSet)) / 2.0f / eps }
       .toNDArray
 
 }
