+++ src/test/scala/fastml4j/classification/LogisticRegressionSuite.scala
@@ -24,17 +24,17 @@ class LogisticRegressionSuite extends FunSuite with BeforeAndAfter  {
 
 
   test("simple synthetic test") {
-    val coef = 3.0
-    val intercept = 1.0
+    val coef = 3.0f
+    val intercept = 1.0f
     val samples = 100
 
     val (points, labels) = generateLogisticInput(intercept, coef, samples, 100)
     val dataSet = new DataSet(points.toNDArray, labels.toNDArray)
     dataSet.validate()
-    val lr = new LogisticRegression(lambdaL2 = 0.0, maxIterations = 10000  ,alpha = 0.1, eps = 1e-10)
+    val lr = new LogisticRegression(lambdaL2 = 0.0f, maxIterations = 10000  ,alpha = 0.1f, eps = 1e-10f)
     lr.fit(dataSet)
-    assert(lr.weights.getDouble(0,0) === coef +- 1)
-    assert(lr.weights.getDouble(0,1) === intercept +- 1)
+    assert(lr.weights.getFloat(0,0) === coef +- 1)
+    assert(lr.weights.getFloat(0,1) === intercept +- 1)
 
   }
